thirdeye/requirements.txt
django-cors-headers
channels
setuptools
djangorestframework_simplejwt
drf-yasg
django
djangorestframework
google-auth
requests
opencv-python
face_recognition
ultralytics
torch
torchvision
numpy
python-decouple
django-notifications-hq
mysqlclient
deep_sort_realtime
asgiref
#django-notifications-hq
